+++ rust/net/infra/src/tcp_ssl.rs
@@ -8,7 +8,6 @@ use std::net::IpAddr;
 use std::time::Duration;
 
 use boring_signal::ssl::{ConnectConfiguration, SslConnector, SslMethod, SslSignatureAlgorithm};
-use tokio::net::TcpStream;
 use tokio_boring_signal::SslStream;
 
 use crate::certs::RootCertificates;
@@ -25,6 +24,11 @@ pub mod proxy;
 
 pub const LONG_TCP_HANDSHAKE_THRESHOLD: Duration = Duration::from_secs(3);
 pub const LONG_TLS_HANDSHAKE_THRESHOLD: Duration = Duration::from_secs(3);
+
+#[cfg(target_os = "macos")]
+pub type TcpStream = crate::stream::WorkaroundWriteBugDuplexStream<tokio::net::TcpStream>;
+#[cfg(not(target_os = "macos"))]
+pub type TcpStream = tokio::net::TcpStream;
 
 #[derive(Clone, Debug)]
 pub struct TcpSslConnector {
@@ -102,9 +106,9 @@ impl Connector<TcpRoute<IpAddr>, ()> for StatelessTcp {
 
         async move {
             let start = tokio::time::Instant::now();
-            tokio::time::timeout(
+            let result = tokio::time::timeout(
                 crate::timeouts::TCP_CONNECTION_TIMEOUT,
-                TcpStream::connect((address, port.get())),
+                tokio::net::TcpStream::connect((address, port.get())),
             )
             .await
             .map_err(|_| {
@@ -112,7 +116,10 @@ impl Connector<TcpRoute<IpAddr>, ()> for StatelessTcp {
                 log::warn!("{log_tag}: TCP connection timed out after {elapsed:?}");
                 TransportConnectError::TcpConnectionFailed
             })?
-            .map_err(|_| TransportConnectError::TcpConnectionFailed)
+            .map_err(|_| TransportConnectError::TcpConnectionFailed)?;
+            #[cfg(target_os = "macos")]
+            let result = crate::stream::WorkaroundWriteBugDuplexStream::new(result);
+            Ok(result)
         }
     }
 }
@@ -233,23 +240,21 @@ pub(crate) mod testutil {
     /// Asserts that the server returns 200 and [`FAKE_RESPONSE`].
     pub(crate) async fn make_http_request_response_over(
         mut stream: impl AsyncRead + AsyncWrite + Unpin,
-    ) {
+    ) -> Result<(), std::io::Error> {
         stream
             .write_all(b"GET /index HTTP/1.1\r\nConnection: close\r\n\r\n")
-            .await
-            .expect("can send request");
+            .await?;
 
         let response = {
             let mut response = String::new();
-            stream
-                .read_to_string(&mut response)
-                .await
-                .expect("receives response");
+            stream.read_to_string(&mut response).await?;
             response
         };
         let lines = response.lines().collect::<Vec<_>>();
 
         assert_eq!(lines.first(), Some("HTTP/1.1 200 OK").as_ref(), "{lines:?}");
         assert_eq!(lines.last(), Some(FAKE_RESPONSE).as_ref(), "{lines:?}");
+
+        Ok(())
     }
 }