+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.76.4
 
+- Work around an issue with misbehaving write(2) calls on macOS.